requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
neo4j>=5.0.0
psycopg2-binary>=2.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
streamlit>=1.40.0
requests>=2.32.0
